The functionality of CurlKiller has been moved into LEPTracking. It is now operated at increasing levels of aggression in order to try to save the event from not not having ANY TPC Tracks, due to too many links in trslnk.
As a first try, all hits are passed to the PATREC. If this fails three
more attempts are made where the size of the bins used to find high
levels of multiplicity in r-phi are increased by factors of 2, 3 and
finally 4. If this still fails an error message is printed warning
that NO TPC Tracks will be reconstructed for this event.

The verbosity of LEPTracking is now controlled completely via
streamlog on the C++ side, and via flags set in tkinit on the F77
side.