Adding processor for PFOAnalysis on TrueJet jets
First commit of this directory: Adding the tool. Combined the
PFOAnalysis processor with the jet analysis power of the
TrueJet/TrueJet_Parser tools. After using TrueJet to find the individual
(true) jets the analysis of PFOAnalysis is applied to give a simple
overview of the basic behaviour of the individual jets.