TrueJet: Reduce the size of the ParticleID vector for final fermion anti fermion pair. Update doc and examples. (#134)
* Reduce the size of the ParticleID vector for the final fermion-antifermion pair, since
for this case, there can only be one pair. This to avoid cluttering of empty collections
after transition to the EDM4HEP world. At the same time, the documentation and example
steerings in the examples subdirectory have been updated. mainly for the move of TrueJet_Parser
from here to MarlinUtil, but also spell-checking etc.
r15 | berggren | 2024-06-10 14:02:15 +0200 (Mon, 10 Jun 2024) | 6 lines

* Modernize and cleanup CMake config for example

* Move CMakeLists for TrueJet to top level directory

* Cleanup and modernize CMake config for TrueJet standalone library

* Fix warnings in example

---------

Co-authored-by: tmadlener <thomas.madlener@desy.de>