Make sure that the digitizer also produces empty output collections (#151)
* Set cellID encoding string whenever possible

Set the encoding as soon as we have the necessary information instead of
waiting until the end as that might be jumped over in case parts of the
loop are skipped.

* Make sure to not jump loop and skip collection registration

We will not enter the loop over the hits in any case if we have 0 hits.
But with the continue in place we jump adding the collection to the
event.